Rename dialect-agnostic classes in coral-hive, coral-spark, coral-trino
Create new Coral-named classes for dialect-agnostic implementations in
coral-hive (12 classes), coral-spark (1 class), and coral-trino (1 class):
- HiveRexBuilder -> CoralRexBuilder
- HiveSqlConformance -> CoralSqlConformance
- HiveSqlToRelConverter -> CoralSqlToRelConverter
- HiveSqlValidator -> CoralSqlValidator
- HiveViewExpander -> CoralViewExpander
- HiveFunction -> CoralFunction
- HiveExplodeOperator -> CoralExplodeOperator
- HiveJsonTupleOperator -> CoralJsonTupleOperator
- HiveNamedStructFunction -> CoralNamedStructFunction
- HivePosExplodeOperator -> CoralPosExplodeOperator
- HiveRLikeOperator -> CoralRLikeOperator
- StaticHiveFunctionRegistry -> StaticCoralFunctionRegistry
- HiveUDFTransformer -> CoralUDFTransformer (in coral-spark and coral-trino)

Original Hive-named classes are retained as @Deprecated thin wrappers
for backward compatibility.
