+++ coral-trino/src/main/java/com/linkedin/coral/trino/rel2trino/transformers/CoralUDFTransformer.java
@@ -1,0 +1,39 @@
+/**
+ * Copyright 2023-2026 LinkedIn Corporation. All rights reserved.
+ * Licensed under the BSD-2 Clause license.
+ * See LICENSE in the project root for license information.
+ */
+package com.linkedin.coral.trino.rel2trino.transformers;
+
+import org.apache.calcite.sql.SqlCall;
+import org.apache.calcite.sql.SqlNodeList;
+import org.apache.calcite.sql.SqlOperator;
+import org.apache.calcite.sql.parser.SqlParserPos;
+
+import com.linkedin.coral.common.transformers.SqlCallTransformer;
+import com.linkedin.coral.hive.hive2rel.functions.VersionedSqlUserDefinedFunction;
+
+
+/**
+ * This transformer converts a UDF SqlCall name from the UDF class name to the
+ * corresponding Trino function name.
+ * i.e. from `com.linkedin.stdudfs.parsing.hive.Ip2Str` to `ip2str`.
+ */
+public class CoralUDFTransformer extends SqlCallTransformer {
+
+  @Override
+  protected boolean condition(SqlCall sqlCall) {
+    final SqlOperator operator = sqlCall.getOperator();
+    final String operatorName = operator.getName();
+    return operator instanceof VersionedSqlUserDefinedFunction && operatorName.contains(".")
+        && !operatorName.equals(".");
+  }
+
+  @Override
+  protected SqlCall transform(SqlCall sqlCall) {
+    final SqlOperator operator = sqlCall.getOperator();
+    final String trinoFunctionName = ((VersionedSqlUserDefinedFunction) operator).getShortFunctionName();
+    return createSqlOperator(trinoFunctionName, operator.getReturnTypeInference())
+        .createCall(new SqlNodeList(sqlCall.getOperandList(), SqlParserPos.ZERO));
+  }
+}
